# docs/swagger/swagger.yaml
basePath: /
definitions:
  blog-api_internal_response.DataEnvelope:
    properties:
      data: {}
      message:
        type: string
    type: object
  blog-api_internal_response.Links:
    properties:
      first:
        type: string
      last:
        type: string
      next:
        type: string
      prev:
        type: string
    type: object
  blog-api_internal_response.MessageEnvelope:
    properties:
      message:
        type: string
    type: object
  blog-api_internal_response.Meta:
    properties:
      current_page:
        type: integer
      filters:
        additionalProperties: {}
        type: object
      from:
        type: integer
      last_page:
        type: integer
      order:
        description: asc | desc
        type: string
      path:
        type: string
      per_page:
        type: integer
      search:
        type: string
      sort:
        type: string
      to:
        type: integer
      total:
        type: integer
    type: object
  blog-api_internal_response.PaginatedEnvelope:
    properties:
      data: {}
      links:
        $ref: '#/definitions/blog-api_internal_response.Links'
      message:
        type: string
      meta:
        $ref: '#/definitions/blog-api_internal_response.Meta'
    type: object
  blog-api_internal_response.ValidationEnvelope:
    properties:
      errors:
        additionalProperties:
          type: string
        type: object
      message:
        type: string
    type: object
  internal_comment.AuthorRef:
    properties:
      public_id:
        type: string
    type: object
  internal_comment.CommentResponse:
    properties:
      author:
        $ref: '#/definitions/internal_comment.AuthorRef'
      body:
        type: string
      created_at:
        type: string
      post:
        $ref: '#/definitions/internal_comment.PostRef'
      public_id:
        type: string
      updated_at:
        type: string
    type: object
  internal_comment.CreateCommentRequest:
    properties:
      body:
        type: string
      post_public_id:
        type: string
    required:
    - body
    - post_public_id
    type: object
  internal_comment.PostRef:
    properties:
      public_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Blog platform API (users, posts, comments). Public identifiers are
    UUID v7 public_id values; raw database ids are never exposed.
  title: Majoo Blog RESTful API
  version: "1.0"
paths:
...
  /api/comments/{publicId}:
    delete:
      parameters:
      - description: Comment public_id (UUID v7)
        in: path
        name: publicId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Delete a comment
      tags:
      - comments
...
  /api/posts/{publicId}/comments:
    get:
      parameters:
      - description: Post public_id (UUID v7)
        in: path
        name: publicId
        required: true
        type: string
      - default: 1
        description: 1-based page index
        in: query
        name: page
        type: integer
      - default: 15
        description: Page size (max 100)
        in: query
        name: per_page
        type: integer
      - description: Free-text search
        in: query
        name: search
        type: string
      - description: Sort column (created_at)
        in: query
        name: sort
        type: string
      - description: Sort direction
        enum:
        - asc
        - desc
        in: query
        name: order
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/blog-api_internal_response.PaginatedEnvelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/internal_comment.CommentResponse'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
      summary: List comments
      tags:
      - comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Post public_id (UUID v7)
        in: path
        name: publicId
        required: true
        type: string
      - description: Comment payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_comment.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/blog-api_internal_response.DataEnvelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_comment.CommentResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/blog-api_internal_response.ValidationEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/blog-api_internal_response.MessageEnvelope'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - comments
...
securityDefinitions:
  BearerAuth:
    description: JWT. Swagger UI uses HTTP bearer (paste the token only).
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"